Update the Spark-Connect test cases to use the decorator function
